Implemented Open_Enter System Call
Implemented a simple callback function which is triggered when a
open_enter system call is detected, and prints out the name of the file
which was opened in the guest. Also renamed the header file to adhere to
PANDA plugin convention.
